Initialize MoveChecker test spec and Converter module
