+++ lib/move_checker.rb
@@ -1,0 +1,4 @@
+#frozen_string_literal: true
+
+class MoveChecker
+end
