+++ spec/move_checker_spec.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative '../lib/move_checker'
+
+RSpec.describe MoveChecker do
+end
